Renamed core folder to cores for clarity
