+++ cpp/src/cores/core_number.cu
