+++ python/cugraph/cores/c_core_number.pxd
