+++ python/cugraph/cores/c_k_core.pxd
